requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
Brotli==1.1.0
cffi==1.16.0
channels==4.1.0
click==8.1.7
constantly==23.10.4
crispy-bootstrap5==0.6
crispy-tailwind==1.0.3
cryptography==42.0.5
daphne==4.1.2
defusedxml==0.7.1
diff-match-patch==20230430
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.2
django-admin-generator==2.6.0
django-appconf==1.0.6
django-auto-logout==0.5.1
django-bootstrap5==24.1
django-cache-url==3.4.5
django-compressor==4.4
django-crispy-forms==2.1
django-datatables-view==1.20.0
django-environ==0.11.2
django-htmx==1.17.3
django-import-export==3.3.7
django-jazzmin==2.6.0
django-online-users==0.3
django-utils2==3.0.2
django-stubs==5.0.0
django-stubs-ext==5.0.0
environs==9.3.5
et-xmlfile==1.1.0
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
MarkupPy==1.14
marshmallow==3.20.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
python-dotenv==1.0.1
python-utils==3.8.2
PyYAML==6.0.1
rcssmin==1.1.1
rjsmin==1.2.1
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2023.4
uvicorn==0.29.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.3
flask==3.0.3
requests==2.32.3
djangorestframework==3.15.2
